[DNM] downcast precision for metrics

+++ tests/metrics/test_preprocesses.py
@@ -46,15 +46,15 @@ def test_preprocess():
     input_tensor = torch.tensor([0.1, 0.5, 0.9])
     target_tensor = torch.tensor([0, 1, 1])
     processed_input, processed_target = preprocess(input_tensor, target_tensor)
-    assert torch.allclose(processed_input, input_tensor)
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_input, input_tensor.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # List inputs
     input_list = [0.1, 0.5, 0.9]
     target_list = [0, 1, 1]
     processed_input, processed_target = preprocess(input_list, target_list)
-    assert torch.allclose(processed_input, torch.tensor(input_list))
-    assert torch.allclose(processed_target, torch.tensor(target_list))
+    assert torch.allclose(processed_input, torch.tensor(input_list).to(processed_input.dtype))
+    assert torch.allclose(processed_target, torch.tensor(target_list).to(processed_target.dtype))
 
     # Ignore index
     input_tensor = torch.tensor([0.1, 0.5, 0.9, 0.3])
@@ -62,8 +62,8 @@ def test_preprocess():
     processed_input, processed_target = preprocess(input_tensor, target_tensor, ignore_index=-100)
     assert processed_input.shape == torch.Size([3])
     assert processed_target.shape == torch.Size([3])
-    assert torch.allclose(processed_input, torch.tensor([0.1, 0.5, 0.3]))
-    assert torch.allclose(processed_target, torch.tensor([0, 1, 0]))
+    assert torch.allclose(processed_input, torch.tensor([0.1, 0.5, 0.3]).to(processed_input.dtype))
+    assert torch.allclose(processed_target, torch.tensor([0, 1, 0]).to(processed_target.dtype))
 
     # Ignore NaN
     input_tensor = torch.tensor([0.1, 0.5, 0.3, 0.1])
@@ -71,20 +71,20 @@ def test_preprocess():
     processed_input, processed_target = preprocess(input_tensor, target_tensor, ignore_nan=True)
     assert processed_input.shape == torch.Size([3])
     assert processed_target.shape == torch.Size([3])
-    assert torch.allclose(processed_input, torch.tensor([0.1, 0.5, 0.3]))
-    assert torch.allclose(processed_target, torch.tensor([0.0, 1.0, 1.0]))
+    assert torch.allclose(processed_input, torch.tensor([0.1, 0.5, 0.3]).to(processed_input.dtype))
+    assert torch.allclose(processed_target, torch.tensor([0.0, 1.0, 1.0]).to(processed_target.dtype))
 
     # NestedTensor
     input_nested = NestedTensor([torch.tensor([0.1, 0.2]), torch.tensor([0.3, 0.4, 0.5])])
     target_nested = NestedTensor([torch.tensor([0, 1]), torch.tensor([1, 0, 1])])
     processed_input, processed_target = preprocess(input_nested, target_nested)
-    assert torch.allclose(processed_input, input_nested.concat)
-    assert torch.allclose(processed_target, target_nested.concat)
+    assert torch.allclose(processed_input, input_nested.concat.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_nested.concat.to(processed_target.dtype))
 
     # Error case: mismatched tensor and nested tensor
     target_tensor = torch.tensor([0, 1, 1, 0, 1])
     with raises(ValueError):
-        processed_input, processed_target = preprocess(input_nested, target_tensor)
+        processed_input, processed_target = preprocess(input_nested, target_tensor.to(processed_input.dtype))
 
     # Nested list
     input_list = [[0.1, 0.2], [0.3, 0.4]]
@@ -99,8 +99,8 @@ def test_preprocess():
     input_batch_tensor = torch.tensor([[0.1, 0.2, 0.0], [0.3, 0.4, 0.5]])
     expected_result = torch.cat([input_batch_tensor[0, :2], input_batch_tensor[1, :3]])
     processed_input, processed_target = preprocess(input_batch_tensor, target_nested)
-    assert torch.allclose(processed_input, expected_result)
-    assert torch.allclose(processed_target, target_nested.concat)
+    assert torch.allclose(processed_input, expected_result.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_nested.concat.to(processed_target.dtype))
 
 
 def test_task_specific_preprocess():
@@ -124,8 +124,8 @@ def test_task_specific_preprocess():
     processed_input, processed_target = preprocess_regression(input_tensor, target_tensor, ignore_nan=True)
     assert processed_input.shape == torch.Size([2])
     assert processed_target.shape == torch.Size([2])
-    assert torch.allclose(processed_input, torch.tensor([0.1, 0.2]))
-    assert torch.allclose(processed_target, torch.tensor([0.1, 0.2]))
+    assert torch.allclose(processed_input, torch.tensor([0.1, 0.2], dtype=processed_input.dtype))
+    assert torch.allclose(processed_target, torch.tensor([0.1, 0.2], dtype=processed_target.dtype))
 
     # Binary classification
     input_tensor = torch.tensor([0.1, 0.7, 0.4])
@@ -133,8 +133,8 @@ def test_task_specific_preprocess():
     processed_input, processed_target = preprocess_binary(input_tensor, target_tensor)
     assert processed_input.shape == torch.Size([3])
     assert processed_target.shape == torch.Size([3])
-    assert torch.allclose(processed_input, input_tensor)
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_input, input_tensor.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # Binary classification with normalization
     input_tensor = torch.tensor([-1.0, 2.0, 0.5])
@@ -143,7 +143,7 @@ def test_task_specific_preprocess():
     assert processed_input.shape == torch.Size([3])
     assert processed_target.shape == torch.Size([3])
     assert processed_input.min() >= 0.0 and processed_input.max() <= 1.0
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # Multiclass classification
     input_tensor = torch.tensor([[0.1, 0.7, 0.2], [0.3, 0.3, 0.4], [0.6, 0.3, 0.1]])
@@ -151,8 +151,8 @@ def test_task_specific_preprocess():
     processed_input, processed_target = preprocess_multiclass(input_tensor, target_tensor, num_classes=3)
     assert processed_input.shape == torch.Size([3, 3])
     assert processed_target.shape == torch.Size([3])
-    assert torch.allclose(processed_input, input_tensor)
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_input, input_tensor.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # Multiclass with normalization
     input_tensor = torch.tensor([[-1.0, 2.0, 0.5], [0.0, -3.0, 1.0], [1.0, 2.0, -1.0]])
@@ -161,8 +161,8 @@ def test_task_specific_preprocess():
     assert processed_input.shape == torch.Size([3, 3])
     assert processed_target.shape == torch.Size([3])
     assert (processed_input >= 0.0).all() and (processed_input <= 1.0).all()
-    assert torch.allclose(processed_input.sum(dim=1), torch.ones(3))
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_input.sum(dim=1), torch.ones(3, dtype=processed_input.dtype))
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # Multilabel classification
     input_tensor = torch.tensor([[0.1, 0.7, 0.2], [0.3, 0.3, 0.4], [0.6, 0.3, 0.1]])
@@ -170,8 +170,8 @@ def test_task_specific_preprocess():
     processed_input, processed_target = preprocess_multilabel(input_tensor, target_tensor, num_labels=3)
     assert processed_input.shape == torch.Size([3, 3])
     assert processed_target.shape == torch.Size([3, 3])
-    assert torch.allclose(processed_input, input_tensor)
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_input, input_tensor.to(processed_input.dtype))
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
 
     # Multilabel with normalization
     input_tensor = torch.tensor([[-1.0, 2.0, 0.5], [0.0, -3.0, 1.0], [1.0, 2.0, -1.0]])
@@ -180,4 +180,4 @@ def test_task_specific_preprocess():
     assert processed_input.shape == torch.Size([3, 3])
     assert processed_target.shape == torch.Size([3, 3])
     assert (processed_input >= 0.0).all() and (processed_input <= 1.0).all()
-    assert torch.allclose(processed_target, target_tensor)
+    assert torch.allclose(processed_target, target_tensor.to(processed_target.dtype))
